backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
django==5.0.7; python_version >= '3.10'
django-cors-headers==4.4.0; python_version >= '3.8'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-spectacular==0.27.2; python_version >= '3.7'
inflection==0.5.1; python_version >= '3.5'
joblib==1.4.2; python_version >= '3.8'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
numpy==2.0.0; python_version >= '3.9'
pandas==2.2.2; python_version >= '3.9'
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.35.1; python_version >= '3.8'
rpds-py==0.19.0; python_version >= '3.8'
scikit-learn==1.5.1; python_version >= '3.9'
scipy==1.14.0; python_version >= '3.10'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.0; python_version >= '3.8'
threadpoolctl==3.5.0; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.11'
tzdata==2024.1; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
